Fix readPreference in QueryBuilder
